# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vault API
  description: >
    Osobní hudební systém — REST API. Pokrývá globální katalog/vyhledávání
    (MusicBrainz/Deezer proxy), asynchronní on-demand provisioning chybějících
    skladeb a doporučení/playlisty (ListenBrainz). Real-time synchronizace
    přehrávání je popsaná odděleně v `asyncapi.yaml`.
  version: "0.1.0"
servers:
  - url: https://vault.local/api/v1
    description: Přes reverse proxy / VPN (Tailscale), nikdy veřejně
security:
  - bearerAuth: []
...
paths:
...
  /catalog/search:
    get:
      tags: [catalog]
      summary: Unifikované vyhledávání napříč lokální DB, MusicBrainz a Deezer
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1 }
        - name: type
          in: query
          schema:
            type: string
            enum: [artist, release, recording]
          description: Omezí výsledky na jeden typ entity; bez parametru se hledá napříč všemi.
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Sjednocené výsledky, každá položka nese `availability`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  query: { type: string }
                  total: { type: integer }
                  results:
                    type: array
                    items: { $ref: "#/components/schemas/SearchResultItem" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /catalog/artists/{artistId}:
    get:
      tags: [catalog]
      summary: Detail interpreta
      parameters:
        - $ref: "#/components/parameters/ArtistId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Artist" }
        "404": { $ref: "#/components/responses/NotFound" }
  /catalog/artists/{artistId}/discography:
    get:
      tags: [catalog]
      summary: Kompletní diskografie interpreta (cache-first z MusicBrainz)
      parameters:
        - $ref: "#/components/parameters/ArtistId"
        - name: releaseType
          in: query
          schema:
            type: string
            enum: [album, ep, single, compilation]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  artist: { $ref: "#/components/schemas/Artist" }
                  releases:
                    type: array
                    items: { $ref: "#/components/schemas/Release" }
        "404": { $ref: "#/components/responses/NotFound" }
  /catalog/releases/{releaseId}:
    get:
      tags: [catalog]
      summary: Detail alba/EP/singlu
      parameters:
        - $ref: "#/components/parameters/ReleaseId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Release" }
        "404": { $ref: "#/components/responses/NotFound" }
  /catalog/releases/{releaseId}/tracks:
    get:
      tags: [catalog]
      summary: Tracklist alba, včetně availability příznaku pro každou skladbu
      parameters:
        - $ref: "#/components/parameters/ReleaseId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Recording" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------------
  # 2. Provisioning — klíčový idempotentní endpoint
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Device-scoped JWT vydaný při párování zařízení.
  parameters:
    ArtistId:
      name: artistId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ReleaseId:
      name: releaseId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Neplatný požadavek
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Chybějící/neplatný token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Zdroj nenalezen
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    Availability:
      type: string
      enum: [available, provisionable, unavailable]
      description: >
        available = MediaAsset.status==AVAILABLE; provisionable = existuje
        v katalogu a lze založit provisioning job; unavailable = žádný
        známý zdroj (čistě metadatová položka bez cesty k obstarání).
    Artist:
      type: object
      properties:
        id: { type: string, format: uuid }
        mbid: { type: string, nullable: true }
        deezerId: { type: string, nullable: true }
        name: { type: string }
        sortName: { type: string }
        images:
          type: array
          items: { type: string, format: uri }
    Release:
      type: object
      properties:
        id: { type: string, format: uuid }
        mbid: { type: string, nullable: true }
        artistId: { type: string, format: uuid }
        title: { type: string }
        releaseDate: { type: string, format: date, nullable: true }
        releaseType: { type: string, enum: [album, ep, single, compilation] }
        images:
          type: array
          items: { type: string, format: uri }
    Recording:
      type: object
      properties:
        id: { type: string, format: uuid }
        mbid: { type: string, nullable: true }
        releaseId: { type: string, format: uuid, nullable: true }
        artistId: { type: string, format: uuid }
        title: { type: string }
        durationMs: { type: integer }
        isrc: { type: string, nullable: true }
        trackNumber: { type: integer, nullable: true }
        availability: { $ref: "#/components/schemas/Availability" }
        previewUrl:
          type: string
          format: uri
          nullable: true
          description: 30s náhled z Deezer (doplňkové metadata, best-effort — nemusí být vyplněno).
        listenCount:
          type: integer
          nullable: true
          description: >
            Počet poslechů z veřejného ListenBrainz API (doplňkové metadata,
            jen u /recommendations/trending a /recommendations/community —
            jinde vždy null).
    SearchResultItem:
      allOf:
        - type: object
          properties:
            entityType: { type: string, enum: [artist, release, recording] }
        - oneOf:
            - $ref: "#/components/schemas/Artist"
            - $ref: "#/components/schemas/Release"
            - $ref: "#/components/schemas/Recording"
...